aadhaar-portal/requirements.txt
# Aadhaar Analytics Portal — Python dependencies
# Install with: pip install -r requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pymysql==1.1.1
cryptography==42.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.7.1
python-dotenv==1.0.1
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.0.0
google-generativeai>=0.7.0
openai>=1.0.0
